Use parallel-webpack
