+++ .gitignore
@@ -7,7 +7,7 @@
 /coverage
 
 # production
-/lib
+/dist
 
 # misc
 .DS_Store
